Tests/LibWeb/Text/expected/css/no-combinators-allowed-in-compound-selector-in-pseudos.txt

::slotted(> a) is invalid: FAIL
::slotted(+ a) is invalid: FAIL
::slotted(~ a) is invalid: FAIL
:host(> a) is invalid: FAIL
:host(+ a) is invalid: FAIL
:host(~ a) is invalid: FAIL
